requirements.txt
fastapi>=0.100.0
pydantic>=1.8.2
uvicorn>=0.15.0
motor>=2.5.0
pydantic
pymongo>=3.11.4
python-dotenv
